fix(compiler): fix parsing error with extension functions
This addresses a bug that lead to a stack overflow exception
when parsing some extension functions.

Specifically, if the text inside the extension function was incomplete (because the user was still typing), we'd end up in a stack overflow.

eg: StockPrices.filterEach(Sto) // where Sto doesn't exist, because it's supposed to be Stock

The root cause was actually a deficiency in how we parse dot-notated stings to symbols. A new SymbolTree has been
added to address this. Over time, we'll likely move more code to use the SymbolTree